Add HomeWork 12-2
