+++ js/main.js
@@ -1,8 +1,9 @@
-import {getServerData} from './server.js';
-import {getPictures} from './get-pictures.js';
+import {getServerData} from './server';
+import {getPictures} from './get-pictures';
 import './form-validate';
 import './photo-scale';
 import './slider-effects';
 import './sort';
+import './photo-preview';
 
 getServerData(getPictures);
